+++ src/promptflow-core/promptflow/core/_errors.py
@@ -90,12 +90,23 @@ class MissingRequiredInputError(CoreError):
     pass
 
 
+class InvalidOutputKeyError(CoreError):
+    """Exception raised when invalid output key."""
+
+    pass
+
+
 class OpenURLUserAuthenticationError(UserAuthenticationError):
     def __init__(self, **kwargs):
         super().__init__(target=ErrorTarget.CORE, **kwargs)
 
 
 class OpenURLFailedUserError(UserErrorException):
+    def __init__(self, **kwargs):
+        super().__init__(target=ErrorTarget.CORE, **kwargs)
+
+
+class OpenURLNotFoundError(UserErrorException):
     def __init__(self, **kwargs):
         super().__init__(target=ErrorTarget.CORE, **kwargs)
 
@@ -124,6 +135,13 @@ class MalformedConnectionProviderConfig(UserErrorException):
             "resourceGroups/<resource_group>/providers/Microsoft.MachineLearningServices/"
             f"workspaces/<workspace_name>, got {provider_config}"
         )
+        super().__init__(target=ErrorTarget.CORE, message=message, **kwargs)
+
+
+class UnsupportedWorkspaceKind(UserErrorException):
+    """Exception raised when workspace kind is not supported."""
+
+    def __init__(self, message, **kwargs):
         super().__init__(target=ErrorTarget.CORE, message=message, **kwargs)